updated + moved match heights function to utils and adjusted all blocks
