+++ _src-lp/blocks/new-prod-boxes/new-prod-boxes.js
@@ -1,5 +1,5 @@
 import { productAliases } from '../../scripts/scripts.js';
-import { updateProductsList } from '../../scripts/utils.js';
+import { updateProductsList, matchHeights } from '../../scripts/utils.js';
 
 function initializeSlider(block) {
   const slidesContainer = block.closest('.slider-container');
@@ -450,49 +450,6 @@ export default function decorate(block) {
       add some products
     </div>`;
   }
-
-  // General function to match the height of elements based on a selector
-  const matchHeights = (targetNode, selector) => {
-    const resetHeights = () => {
-      const elements = targetNode.querySelectorAll(selector);
-      elements.forEach((element) => {
-        element.style.minHeight = '';
-      });
-    };
-
-    const adjustHeights = () => {
-      if (window.innerWidth >= 768) {
-        resetHeights();
-        const elements = targetNode.querySelectorAll(selector);
-        const elementsHeight = Array.from(elements).map((element) => element.offsetHeight);
-        const maxHeight = Math.max(...elementsHeight);
-
-        elements.forEach((element) => {
-          element.style.minHeight = `${maxHeight}px`;
-        });
-      } else {
-        resetHeights();
-      }
-    };
-
-    const matchHeightsCallback = (mutationsList) => {
-      Array.from(mutationsList).forEach((mutation) => {
-        if (mutation.type === 'childList') {
-          adjustHeights();
-        }
-      });
-    };
-
-    const observer = new MutationObserver(matchHeightsCallback);
-
-    if (targetNode) {
-      observer.observe(targetNode, { childList: true, subtree: true });
-    }
-
-    window.addEventListener('resize', () => {
-      adjustHeights();
-    });
-  };
 
   const targetNode = document.querySelector('.new-prod-boxes');
   matchHeights(targetNode, '.tag-subtitle');
