+++ _src-lp/scripts/utils.js
@@ -786,3 +786,71 @@ export function openUrlForOs(urlMacos, urlWindows, urlAndroid, urlIos) {
     window.open(openUrl, '_self');
   }
 }
+
+// General function to match the height of elements based on a selector
+export function debounce(func, wait) {
+  let timeout;
+  return function executedFunction(...args) {
+    const later = () => {
+      clearTimeout(timeout);
+      func(...args);
+    };
+    clearTimeout(timeout);
+    timeout = setTimeout(later, wait);
+  };
+}
+
+export async function matchHeights(targetNode, selector) {
+  const resetHeights = () => {
+    const elements = targetNode.querySelectorAll(selector);
+    elements.forEach((element) => {
+      element.style.minHeight = '';
+    });
+  };
+
+  const adjustHeights = () => {
+    if (window.innerWidth >= 768) {
+      resetHeights();
+      const elements = targetNode.querySelectorAll(selector);
+      const elementsHeight = Array.from(elements).map((element) => element.offsetHeight);
+      const maxHeight = Math.max(...elementsHeight);
+
+      elements.forEach((element) => {
+        element.style.minHeight = `${maxHeight}px`;
+      });
+    } else {
+      resetHeights();
+    }
+  };
+
+  const matchHeightsCallback = (mutationsList) => {
+    Array.from(mutationsList).forEach((mutation) => {
+      if (mutation.type === 'childList') {
+        adjustHeights();
+      }
+    });
+  };
+
+  const observer = new MutationObserver(matchHeightsCallback);
+  const resizeObserver = new ResizeObserver(debounce((entries) => {
+    // eslint-disable-next-line no-unused-vars
+    entries.forEach((entry) => {
+      adjustHeights();
+    });
+  }), 100);
+
+  if (targetNode) {
+    observer.observe(targetNode, { childList: true, subtree: true });
+  }
+
+  window.addEventListener('resize', () => {
+    adjustHeights();
+  });
+
+  const elements = targetNode.querySelectorAll(selector);
+  elements.forEach((element) => {
+    resizeObserver.observe(element);
+  });
+
+  adjustHeights();
+}
